requirements.txt
Django==2.1.2
django-extensions==2.1.3
django-filter==2.0.0
django-webpack-loader==0.6.0
djangorestframework==3.9.0
Markdown==3.0.1
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pudb==2018.1
Pygments==2.2.0
pytz==2018.5
six==1.11.0
urwid==2.0.1
